# Optimize denotation resolution