This is not really faster...

```sh
> hyperfine -w 5 'target/release/stak ~/baz.scm' '~/src/github.com/raviqqe/stak/target/release/stak ~/baz.scm'
Benchmark 1: target/release/stak ~/baz.scm
  Time (mean ± σ):      1.071 s ±  0.002 s    [User: 1.065 s, System: 0.001 s]
  Range (min … max):    1.066 s …  1.075 s    10 runs

Benchmark 2: ~/src/github.com/raviqqe/stak/target/release/stak ~/baz.scm
  Time (mean ± σ):      1.071 s ±  0.003 s    [User: 1.062 s, System: 0.001 s]
  Range (min … max):    1.068 s …  1.075 s    10 runs

Summary
  ~/src/github.com/raviqqe/stak/target/release/stak ~/baz.scm ran
    1.00 ± 0.00 times faster than target/release/stak ~/baz.scm
```